repositories/voice-agent/pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Voice AI agent for Leonobitech - LiveKit + Whisper + Piper + Claude"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents[anthropic,silero,turn-detector,noise-cancellation,mcp]~=1.3",
    "faster-whisper",
    "piper-tts",
    "numpy",
    "python-dotenv",
]

[project.scripts]
start = "agent:server"
